fix ef
