+++ src/analysis/retail/monk/mistweaver/CombatLogParser.ts
@@ -37,7 +37,6 @@ import CloudedFocus from './modules/spells/CloudedFocus';
 import EnvelopingBreath from './modules/spells/EnvelopingBreath';
 import EnvelopingMists from './modules/spells/EnvelopingMists';
 import EssenceFont from './modules/spells/EssenceFont';
-import EssenceFontCancelled from './modules/spells/EssenceFontCancelled';
 import EssenceFontTargetsHit from './modules/spells/EssenceFontTargetsHit';
 import EssenceFontUniqueTargets from './modules/spells/EssenceFontUniqueTargets';
 import ExpelHarm from './modules/spells/ExpelHarm';
@@ -140,7 +139,6 @@ class CombatLogParser extends CoreCombatLogParser {
     essenceFontUniqueTargets: EssenceFontUniqueTargets,
     essenceFontTargetsHit: EssenceFontTargetsHit,
     expelHarm: ExpelHarm,
-    EssenceFontCancelled: EssenceFontCancelled,
     faelineStomp: FaelineStomp,
     faelineStompHealing: FaelineStompHealing,
     invokersDelight: InvokersDelight,
