+++ src/analysis/retail/monk/mistweaver/modules/features/Checklist/Module.tsx
@@ -3,9 +3,7 @@ import CastEfficiency from 'parser/shared/modules/CastEfficiency';
 import Combatants from 'parser/shared/modules/Combatants';
 import BaseChecklist from 'parser/shared/modules/features/Checklist/Module';
 import ManaValues from 'parser/shared/modules/ManaValues';
-
 import EnvelopingBreath from '../../spells/EnvelopingBreath';
-import EssenceFontCancelled from '../../spells/EssenceFontCancelled';
 import EssenceFontTargetsHit from '../../spells/EssenceFontTargetsHit';
 import SoothingMist from '../../spells/SoothingMist';
 import SpinningCraneKick from '../../spells/SpinningCraneKick';
@@ -23,6 +21,7 @@ import Component from './Component';
 import VivaciousVivification from '../../spells/VivaciousVivify';
 import AncientTeachingsoftheMonastery from '../../spells/AncientTeachingsoftheMonastery';
 import SheilunsGift from '../../spells/SheilunsGift';
+import EssenceFont from '../../spells/EssenceFont';
 
 class Checklist extends BaseChecklist {
   static dependencies = {
@@ -45,7 +44,7 @@ class Checklist extends BaseChecklist {
     jadeSerpentStatue: JadeSerpentStatue,
     soothingMist: SoothingMist,
     envelopingBreath: EnvelopingBreath,
-    EssenceFontCancelled: EssenceFontCancelled,
+    essenceFont: EssenceFont,
     vivaciousVivification: VivaciousVivification,
     ancientTeachingsoftheMonastery: AncientTeachingsoftheMonastery,
     sheiluns: SheilunsGift,
@@ -68,7 +67,7 @@ class Checklist extends BaseChecklist {
   protected jadeSerpentStatue!: JadeSerpentStatue;
   protected soothingMist!: SoothingMist;
   protected envelopingBreath!: EnvelopingBreath;
-  protected EssenceFontCancelled!: EssenceFontCancelled;
+  protected essenceFont!: EssenceFont;
   protected vivaciousVivification!: VivaciousVivification;
   protected ancientTeachingsoftheMonastery!: AncientTeachingsoftheMonastery;
   protected sheiluns!: SheilunsGift;
@@ -98,7 +97,7 @@ class Checklist extends BaseChecklist {
           vivify: this.vivify.suggestionThresholds,
           jadeSerpentStatue: this.jadeSerpentStatue.suggestionThresholds,
           soothingMist: this.soothingMist.suggestionThresholdsCasting,
-          EssenceFontCancelled: this.EssenceFontCancelled.suggestionThresholds,
+          essenceFontCancelled: this.essenceFont.suggestionThresholds,
           vivaciousVivification: this.vivaciousVivification.suggestionThresholds,
           ancientTeachings: this.ancientTeachingsoftheMonastery.suggestionThresholds,
           sheiluns: this.sheiluns.suggestionThresholds,
